Rename expressions file
